chore: rspack test tools diff demo
