+++ packages/rspack-test-tools/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @rspack/test-tools
