+++ packages/rspack-test-tools/README.md
@@ -1,0 +1,16 @@
+<picture>
+  <source media="(prefers-color-scheme: dark)" srcset="https://lf3-static.bytednsdoc.com/obj/eden-cn/rjhwzy/ljhwZthlaukjlkulzlp/rspack-banner-1610-dark.png">
+  <img alt="Rspack Banner" src="https://lf3-static.bytednsdoc.com/obj/eden-cn/rjhwzy/ljhwZthlaukjlkulzlp/rspack-banner-1610.png">
+</picture>
+
+# @rspack/test-tools
+
+Test tools for rspack.
+
+## Documentation
+
+See [https://rspack.dev](https://rspack.dev) for details.
+
+## License
+
+Rspack is [MIT licensed](https://github.com/web-infra-dev/rspack/blob/main/LICENSE).
