+++ packages/rspack-test-tools/package.json
@@ -1,0 +1,60 @@
+{
+  "name": "@rspack/test-tools",
+  "version": "0.3.12",
+  "license": "MIT",
+  "description": "Test tools for rspack",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "default": "./dist/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "build": "rimraf dist/ && tsc -b ./tsconfig.build.json --force",
+    "dev": "tsc -b -w",
+    "test": "jest --runInBand"
+  },
+  "files": [
+    "client",
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "homepage": "https://rspack.dev",
+  "bugs": "https://github.com/web-infra-dev/rspack/issues",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/web-infra-dev/rspack",
+    "directory": "packages/rspack-test-tools"
+  },
+  "dependencies": {
+    "@babel/generator": "7.23.0",
+    "@babel/helpers": "7.23.2",
+    "@babel/parser": "7.23.0",
+    "@babel/template": "7.22.15",
+    "@babel/traverse": "7.23.2",
+    "@babel/types": "7.23.0",
+    "@rspack/core": "workspace:*",
+    "deepmerge": "^4.3.1",
+    "fs-extra": "^11.1.1",
+    "jest-diff": "^29.7.0",
+    "webpack-sources": "3.2.3"
+  },
+  "devDependencies": {
+    "@types/webpack": "5.28.3",
+    "@types/webpack-sources": "3.2.0",
+    "webpack": "5.89.0"
+  },
+  "peerDependencies": {},
+  "peerDependenciesMeta": {},
+  "jest": {
+    "watchPathIgnorePatterns": [
+      "<rootDir>/dist",
+      "<rootDir>/tests/dist"
+    ],
+    "testEnvironment": "../../scripts/test/patch-node-env.cjs"
+  }
+}
