+++ packages/rspack-test-tools/src/compare/index.ts
@@ -1,0 +1,4 @@
+export * from "./replace-module-argument";
+export * from "./replace-runtime-module-name";
+export * from "./format-code";
+export * from "./compare";
