+++ packages/rspack-test-tools/src/compare/replace-module-argument.ts
@@ -1,0 +1,3 @@
+export function replaceModuleArgument(raw: string) {
+	return raw.trim().replace(/^\(function \([\w_,\s]+\) {/, "(function () {");
+}
