+++ packages/rspack-test-tools/src/helper/parse-modules.ts
