+++ packages/rspack-test-tools/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./type";
+export * from "./test/tester";
+export * from "./processor/diff";
