+++ packages/rspack-test-tools/src/webpack/module-placeholder-plugin.ts
@@ -1,63 +1,70 @@
-const JavascriptModulesPlugin = require("webpack/lib/javascript/JavascriptModulesPlugin");
+// @ts-nocheck
 const {
 	ConcatSource,
 	RawSource,
 	CachedSource,
 	PrefixSource
 } = require("webpack-sources");
-const Template = require("webpack/lib/Template");
+const path = require("path");
 
-function renderRuntimeModules(runtimeModules, renderContext) {
-	const source = new ConcatSource();
-	for (const module of runtimeModules) {
-		const codeGenerationResults = renderContext.codeGenerationResults;
-		let runtimeSource;
-		if (codeGenerationResults) {
-			runtimeSource = codeGenerationResults.getSource(
-				module,
-				renderContext.chunk.runtime,
-				"runtime"
-			);
-		} else {
-			const codeGenResult = module.codeGeneration({
-				chunkGraph: renderContext.chunkGraph,
-				dependencyTemplates: renderContext.dependencyTemplates,
-				moduleGraph: renderContext.moduleGraph,
-				runtimeTemplate: renderContext.runtimeTemplate,
-				runtime: renderContext.chunk.runtime,
-				codeGenerationResults
-			});
-			if (!codeGenResult) continue;
-			runtimeSource = codeGenResult.sources.get("runtime");
-		}
-		if (runtimeSource) {
-			source.add(
-				Template.toNormalComment(`start::${module.identifier()}`) + "\n"
-			);
-			if (!module.shouldIsolate()) {
-				source.add(runtimeSource);
-				source.add("\n\n");
-			} else if (renderContext.runtimeTemplate.supportsArrowFunction()) {
-				source.add("(() => {\n");
-				source.add(new PrefixSource("\t", runtimeSource));
-				source.add("\n})();\n\n");
+function createRenderRuntimeModulesFn(Template) {
+	return function renderRuntimeModules(runtimeModules, renderContext) {
+		const source = new ConcatSource();
+		for (const module of runtimeModules) {
+			const codeGenerationResults = renderContext.codeGenerationResults;
+			let runtimeSource;
+			if (codeGenerationResults) {
+				runtimeSource = codeGenerationResults.getSource(
+					module,
+					renderContext.chunk.runtime,
+					"runtime"
+				);
 			} else {
-				source.add("!function() {\n");
-				source.add(new PrefixSource("\t", runtimeSource));
-				source.add("\n}();\n\n");
+				const codeGenResult = module.codeGeneration({
+					chunkGraph: renderContext.chunkGraph,
+					dependencyTemplates: renderContext.dependencyTemplates,
+					moduleGraph: renderContext.moduleGraph,
+					runtimeTemplate: renderContext.runtimeTemplate,
+					runtime: renderContext.chunk.runtime,
+					codeGenerationResults
+				});
+				if (!codeGenResult) continue;
+				runtimeSource = codeGenResult.sources.get("runtime");
 			}
-			source.add(
-				Template.toNormalComment(`end::${module.identifier()}`) + "\n"
-			);
+			if (runtimeSource) {
+				source.add(
+					Template.toNormalComment(`start::${module.identifier()}`) + "\n"
+				);
+				if (!module.shouldIsolate()) {
+					source.add(runtimeSource);
+					source.add("\n\n");
+				} else if (renderContext.runtimeTemplate.supportsArrowFunction()) {
+					source.add("(() => {\n");
+					source.add(new PrefixSource("\t", runtimeSource));
+					source.add("\n})();\n\n");
+				} else {
+					source.add("!function() {\n");
+					source.add(new PrefixSource("\t", runtimeSource));
+					source.add("\n}();\n\n");
+				}
+				source.add(
+					Template.toNormalComment(`end::${module.identifier()}`) + "\n"
+				);
+			}
 		}
-	}
-	return source;
+		return source;
+	};
 }
 
 const caches = new WeakMap();
 
-export function createModulePlaceholderPlugin() {
-	Template.renderRuntimeModules = renderRuntimeModules;
+export function createModulePlaceholderPlugin(webpackPath) {
+	const JavascriptModulesPlugin = require(path.join(
+		path.dirname(webpackPath),
+		"javascript/JavascriptModulesPlugin.js"
+	));
+	const Template = require(path.join(path.dirname(webpackPath), "Template.js"));
+	Template.renderRuntimeModules = createRenderRuntimeModulesFn(Template);
 	return {
 		apply(compiler) {
 			compiler.hooks.compilation.tap("RuntimeDiffPlugin", compilation => {
