+++ packages/rspack/package.json
@@ -18,7 +18,9 @@
     "test": "cross-env NO_COLOR=1 RSPACK_DEP_WARNINGS=false node --expose-gc --max-old-space-size=8192 --experimental-vm-modules ../../node_modules/jest-cli/bin/jest --runInBand --logHeapUsage --testPathIgnorePatterns \".diff.test.ts\"",
     "test-diff": "cross-env RSPACK_DIFF=true NO_COLOR=1 RSPACK_DEP_WARNINGS=false node --expose-gc --max-old-space-size=8192 --experimental-vm-modules ../../node_modules/jest-cli/bin/jest --runInBand --logHeapUsage --testPathPattern \".diff.test.ts\""
   },
-  "files": ["dist"],
+  "files": [
+    "dist"
+  ],
   "engines": {
     "node": ">=16.0.0"
   },
@@ -35,19 +37,19 @@
     "@rspack/plugin-node-polyfill": "workspace:^",
     "@swc/core": "^1.3.96",
     "@swc/jest": "^0.2.29",
+    "@rspack/test-tools": "workspace:^",
     "@types/watchpack": "^2.4.0",
     "@types/webpack": "5.28.3",
+    "@types/webpack-dev-server": "^4.7.2",
     "@types/webpack-sources": "3.2.0",
     "@types/ws": "8.5.3",
     "babel-loader": "^9.1.0",
     "babel-plugin-import": "^1.13.5",
     "copy-webpack-plugin": "5.1.2",
     "cross-env": "^7.0.3",
-    "deepmerge": "^4.3.1",
     "file-loader": "^6.2.0",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest-diff": "^29.7.0",
     "jest-serializer-path": "^0.1.15",
     "less": "4.1.3",
     "less-loader": "^11.1.0",
@@ -61,8 +63,8 @@
     "styled-components": "^6.0.8",
     "terser": "5.16.1",
     "ts-node": "10.9.1",
-    "wast-loader": "^1.11.4",
-    "webpack": "5.89.0"
+    "webpack": "5.89.0",
+    "wast-loader": "^1.11.4"
   },
   "dependencies": {
     "@rspack/binding": "workspace:*",
