--- packages/rspack/tests/helpers/tester/case/index.ts
@@ -1,1 +1,0 @@
-export * from "./diff";
