--- packages/rspack/tests/helpers/tester/helper/index.ts
@@ -1,7 +1,0 @@
-export * from "./read-config-file";
-export * from "./run-build";
-export * from "./parse-modules";
-export * from "./replace-module-argument";
-export * from "./replace-runtime-module-name";
-export * from "./format-code";
-export * from "./compare-modules";
