--- packages/rspack/tests/helpers/tester/helper/replace-module-argument.ts
@@ -1,3 +1,0 @@
-export function replaceModuleArgument(raw: string) {
-	return raw.trim().replace(/^\(function \([\w_\,\s]+\) {/, "(function () {");
-}
