--- packages/rspack/tests/helpers/tester/index.ts
@@ -1,4 +1,0 @@
-export * from "./type";
-export * from "./tester";
-export * from "./processor";
-export * from "./case";
