--- packages/rspack/tests/helpers/tester/processor/index.ts
@@ -1,1 +1,0 @@
-export * from "./diff";
